Changed to use memcpy
Changed to use memcpy, which allows for the data input in the stack_push to have a void datatype. This might solve the problem of not being able to change the type of data that is pushed to the stack. A self contained solution could be implemented with a for loop that copies memory byte by byte, however as an initial step we will see if memcpy solves the problems.