Code corrections
The stack was always allocating 10 elements, regardless of stack_size_allocated, changed that. Also, the if condition to reallocate bits of the stack (whenever it needed to be increased) had a legacy code portion based on byte size allocation in stack_size_allocated, which has been changed to number of elements. Also, it was allocating one more than needed (>= -> >). Also tested the code for different types of input elements, and it seems to be working.