Updated simplemain.c
Added three stress tests to the stack. One for high memory usage, with a considerable stack size (around 0.7 GB of RAM needed), and another for high CPU usage, two stacks to which elements get added and removed multiple times (the size never increases, but there are a lot of calls to the push and pop functions). Finally, another to test how many stacks it can handle was also developed (tested for creating 20000 stacks).